+++ ChangeLog.md
@@ -20,6 +20,7 @@
     - Define test/sh for services.pl vars (todo #129)
     - Improve test/sh/t-daemon-vars, t-dbms-vars, t-ttp-vars to test key overridings (todo #150)
     - Let test/cmd/run.cmd accept a list of tests as command-line arguments
+    - Improve test/cmd/t-daemon-vars, t-dbms-vars, t-ttp-vars to test key overridings (todo #150)
 
 ### 4.13.0